server/requirements.txt
amqp==5.1.1
auth0-python==4.1.0
Authlib==1.2.0
azure-ai-formrecognizer==3.2.1
azure-common==1.1.28
azure-core==1.26.3
azure-storage-blob==12.15.0
billiard==3.6.4.0
camel-converter==3.0.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==40.0.1
distlib==0.3.6
dnspython==2.3.0
email-validator==1.3.1
eventlet==0.33.3
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
meilisearch==0.25.0
mongoengine==0.27.0
msrest==0.7.1
oauthlib==3.2.2
pipenv==2023.3.20
platformdirs==3.2.0
prompt-toolkit==3.0.38
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
Werkzeug==2.2.3
WTForms==3.0.1
